add CMake support
this commit adds Cmake support that coexists with the
existing Makefiles and regression tests. In order not to trash
the existing Makefiles an out-of-source build is required.
Create an empty directory (can be in the source tree) cd in,
and configure and build from there.

CMake build add a couple of useful features such as automatically
locating third party dependencies (zlib) in locations other than
/usr/lib, adding a configurable install, and enabling parallel
builds. The regression tests have been ported, the way to run them
is from the build dir "make -j4  && make test" as test target
doesn't depend on the test executables (an unfortunate feature of
CMake).