Fix including xlsxwriter as a CocoaPod on macOS.
Explanation for this change:

Previously, CocoaPods would set up the framework target such that `xlsxwriter.h` would appear in the framework's main directory, alongside the `xlsxwriter` binary. It appears that the `codesign` utility only expects binaries in that directory, though, with all headers being in the `Headers` subdirectory (otherwise it fails with an `xlsxwriter.h: code object is not signed at all` message). This change modifies the header_mappings_dir to ensure that `xlsxwriter.h` is in `Headers`, with all other headers in `Headers/xlsxwriter`, and adjusts the paths in the module map accordingly.

In addition, the umbrella header is modified to include Foundation rather than UIKit, as the former is sufficient to compile the module while the latter is not available on macOS.