Use DESTDIR and PREFIX instead of INSTALL_LOCATION
PREFIX is the final location on the user's system. DESTDIR could
be staging directory for a package build, a chroot, etc. The
distinction is important because you don't want DESTDIR to end up
in paths compiled into the program, e.g. the location of a helpfile
or library.

(Applies only to the makefiles, not CMake)

Resolves #216